pyproject.toml
[tool.poetry]
name = "rag-time-workshop"
version = "0.1.0"
description = ""
authors = ["Fabrizio Mele <fabrizio@mele.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
langchain = "^0.2.3"
langchain-aws = "^0.1.6"
langchain-community = "^0.2.4"
langchain-text-splitters = "^0.2.1"
tiktoken = "^0.7.0"
faiss-cpu = "^1.8.0"
boto3 = "^1.34.123"
python-dotenv = "^1.0.1"
pymupdf = "^1.24.5"
bs4 = "^0.0.2"
asyncio = "^3.4.3"
langchain-chroma = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
